+++ src/main/java/org/elsys/fatcatserver/FatcatServerRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FatcatServerRepository extends JpaRepository<Person, Integer> {
+public interface FatcatServerRepository extends JpaRepository<Person, Long> {
 
 }
